Skip font install if already available

+++ install-font.sh
@@ -1,21 +1,28 @@
 #!/bin/bash
 
-# Create a temporary directory
-TEMP_DIR=$(mktemp -d)
+if [ -z "$(fc-list : family  sort | uniq | grep 'JetBrainsMono Nerd Font')" ]; then
+    echo "No fonts found. Running install steps..."
+    # Create a temporary directory
+    TEMP_DIR=$(mktemp -d)
+    
+    # Download the font zip file
+    wget -O "$TEMP_DIR/font.zip" "https://github.com/ryanoasis/nerd-fonts/releases/download/v3.4.0/JetBrainsMono.zip"
+    
+    # Unzip the font file
+    unzip "$TEMP_DIR/font.zip" -d "$TEMP_DIR"
+    
+    # Move the font files to the system fonts directory
+    sudo mv "$TEMP_DIR"/*.{ttf,otf} /usr/share/fonts/
+    
+    # Update the font cache
+    fc-cache -f -v
+    
+    # Clean up
+    rm -rf "$TEMP_DIR"
+    
+    echo "Fonts installed successfully!"
+else
+    echo "Fonts are already installed."
+fi
 
-# Download the font zip file
-wget -O "$TEMP_DIR/font.zip" "https://github.com/ryanoasis/nerd-fonts/releases/download/v3.4.0/JetBrainsMono.zip"
 
-# Unzip the font file
-unzip "$TEMP_DIR/font.zip" -d "$TEMP_DIR"
-
-# Move the font files to the system fonts directory
-sudo mv "$TEMP_DIR"/*.{ttf,otf} /usr/share/fonts/
-
-# Update the font cache
-fc-cache -f -v
-
-# Clean up
-rm -rf "$TEMP_DIR"
-
-echo "Fonts installed successfully!"
